request code separated
remote :Int for error type enums
